refactor: separate concerns between gradient descent and logistic regression
