+++ src/test/java/org/rsultan/regression/LinearRegressionTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 import org.nd4j.linalg.api.buffer.DataType;
 import org.nd4j.linalg.factory.Nd4j;
 import org.rsultan.dataframe.Dataframes;
+import org.rsultan.regression.impl.LinearRegression;
 import org.rsultan.utils.CSVUtilsTest;
 
 import java.io.File;
