+++ src/test/java/org/rsultan/regression/LogisticRegressionTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 import org.nd4j.linalg.api.buffer.DataType;
 import org.nd4j.linalg.factory.Nd4j;
 import org.rsultan.dataframe.Dataframes;
+import org.rsultan.regression.impl.LogisticRegression;
 import org.rsultan.utils.CSVUtilsTest;
 
 import java.io.File;
@@ -56,9 +57,9 @@ public class LogisticRegressionTest {
                 .setPredictorNames(predictors)
                 .setResponseVariableName(responseVariable)
                 .setPredictionColumnName("predictions")
-                .setLabel(label)
-                .train(dataframe);
-        logisticRegression.setLossAccuracyOffset(1);
+                .setLabel(label);
+        logisticRegression.setLossAccuracyOffset(10);
+        logisticRegression.train(dataframe);
         logisticRegression.getHistory().tail();
 
         var dfPredict = logisticRegression.predict(dataframe);
